tools/tools/requirements.txt
requests>=2.32.4
jira>=3.5.1
slack-sdk>=3.26.1
python-dotenv>=1.0.0
